co-located diff data
