+++ frontend/src/tests/import-export-data/diff-data.js
@@ -1,8 +1,9 @@
 /*
- * A collection of initial objects
- * used for computing difference with modified objects.
+ * A collection of initial objects and modified objects
+ * used for testing diff functionalities
  */
 
+// instructor data to compare and compute diff
 export const initialInstructorData = {
     instructors: [
         {
@@ -19,6 +20,27 @@ export const initialInstructorData = {
     ],
 };
 
+export const modifiedInstructorData = [
+    {
+        email: "hery.smith@utoronto.ca",
+        first_name: "Henry",
+        last_name: "Smith",
+        utorid: "smithh",
+    },
+    {
+        email: "gordon.smith@utoronto.ca",
+        first_name: "戈登",
+        utorid: "smithhg",
+    },
+    {
+        email: "megan.miller@utoronto.ca",
+        first_name: "Miller",
+        last_name: "Miles",
+        utorid: "millerm",
+    },
+];
+
+// applicant data to compare and compute diff
 export const initialApplicantData = {
     applicants: [
         {
@@ -39,6 +61,34 @@ export const initialApplicantData = {
     ],
 };
 
+export const modifiedApplicantData = [
+    {
+        email: "goofy-duck@donald.com",
+        first_name: "John",
+        last_name: "Doe",
+        phone: "4166666666",
+        student_number: "10000000",
+        utorid: "johnd",
+    },
+    {
+        email: undefined,
+        first_name: "Ron",
+        last_name: "Weasley",
+        phone: undefined,
+        student_number: undefined,
+        utorid: "weasleyr",
+    },
+    {
+        email: "harry@potter.com",
+        first_name: "哈利",
+        last_name: "Potter",
+        phone: "41888888888",
+        student_number: "999666999",
+        utorid: "potterh",
+    },
+];
+
+// position data to compare and compute diff
 export const initialPositionData = {
     positions: [
         {
@@ -109,6 +159,69 @@ export const initialPositionData = {
     ],
 };
 
+export const modifiedPositionData = [
+    {
+        contract_template: "Regular",
+        current_enrollment: undefined,
+        current_waitlisted: undefined,
+        desired_num_assignments: undefined,
+        duties: "",
+        end_date: "2020-05-01T00:00:00.000Z",
+        hours_per_assignment: 20,
+        instructors: "",
+        position_code: "CSC494",
+        position_title: "Capstone Project",
+        qualifications: "",
+        start_date: "2020-01-01T00:00:00.000Z",
+    },
+    {
+        contract_template: "Regular",
+        current_enrollment: undefined,
+        current_waitlisted: undefined,
+        desired_num_assignments: undefined,
+        duties: "",
+        end_date: "2020-12-31T00:00:00.000Z",
+        hours_per_assignment: 70,
+        instructors: "Garcia, Emily; Smith, Henry",
+        position_code: "MAT136H1F",
+        position_title: "代数",
+        qualifications: "",
+        start_date: "2020-02-10T00:00:00.000Z",
+    },
+    {
+        contract_template: "Regular",
+        current_enrollment: undefined,
+        current_waitlisted: undefined,
+        desired_num_assignments: undefined,
+        duties: "Tutorials",
+        end_date: "2020-12-31T00:00:00.000Z",
+        hours_per_assignment: 75,
+        instructors: "Smith, Henry",
+        position_code: "CSC135H1F",
+        position_title: "Computer Fun",
+        qualifications: "",
+        start_date: "2020-02-10T00:00:00.000Z",
+    },
+];
+
+export const modifiedPositionDataInvalidInstructor = [
+    {
+        contract_template: "Regular",
+        current_enrollment: undefined,
+        current_waitlisted: undefined,
+        desired_num_assignments: undefined,
+        duties: "",
+        end_date: "2020-12-31T00:00:00.000Z",
+        hours_per_assignment: 70,
+        instructors: "Invalid, Instructor",
+        position_code: "MAT136H1F",
+        position_title: "代数",
+        qualifications: "",
+        start_date: "2020-02-10T00:00:00.000Z",
+    },
+];
+
+// assignment data to compare and compute diff
 export const initialAssignmentData = {
     assignments: [
         {
@@ -258,6 +371,161 @@ export const initialAssignmentData = {
     session: { rate: 50, rate1: 50, rate2: 50, rate3: 50 },
 };
 
+export const modifiedAssignmentData = [
+    {
+        contract_override_pdf: null,
+        contract_template: "regular",
+        end_date: "2021-12-10T00:00:00.000Z",
+        hours: 80,
+        position_code: "CSC494",
+        start_date: "2020-12-10T00:00:00.000Z",
+        utorid: "potterh",
+        wage_chunks: [
+            {
+                end_date: "2020-12-31T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2020-12-10T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-06-30T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2021-01-01T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-12-10T00:00:00.000Z",
+                hours: 20,
+                rate: 50,
+                start_date: "2021-07-01T00:00:00.000Z",
+            },
+        ],
+    },
+    {
+        contract_override_pdf: null,
+        contract_template: "regular",
+        end_date: "2021-12-10T00:00:00.000Z",
+        hours: 100,
+        position_code: "CSC494",
+        start_date: "2020-12-10T00:00:00.000Z",
+        utorid: "weasleyr",
+        wage_chunks: [
+            {
+                end_date: "2020-12-31T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2020-12-10T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-06-30T00:00:00.000Z",
+                hours: 10,
+                rate: 50,
+                start_date: "2021-01-01T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-12-10T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2021-07-01T00:00:00.000Z",
+            },
+        ],
+    },
+    {
+        contract_override_pdf: null,
+        contract_template: "regular",
+        end_date: "2021-12-10T00:00:00.000Z",
+        hours: 80,
+        position_code: "CSC494",
+        start_date: "2020-12-10T00:00:00.000Z",
+        utorid: "doej",
+        wage_chunks: [
+            {
+                end_date: "2020-12-31T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2020-12-10T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-06-30T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2021-01-01T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-12-10T00:00:00.000Z",
+                hours: 20,
+                rate: 50,
+                start_date: "2021-07-01T00:00:00.000Z",
+            },
+        ],
+    },
+];
+
+export const modifiedAssignmentDataInvalidApplicant = [
+    {
+        contract_override_pdf: null,
+        contract_template: "regular",
+        end_date: "2021-12-10T00:00:00.000Z",
+        hours: 80,
+        position_code: "CSC494",
+        start_date: "2020-12-10T00:00:00.000Z",
+        utorid: "invalidapplicant",
+        wage_chunks: [
+            {
+                end_date: "2020-12-31T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2020-12-10T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-06-30T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2021-01-01T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-12-10T00:00:00.000Z",
+                hours: 20,
+                rate: 50,
+                start_date: "2021-07-01T00:00:00.000Z",
+            },
+        ],
+    },
+];
+
+export const modifiedAssignmentDataInvalidPosition = [
+    {
+        contract_override_pdf: null,
+        contract_template: "regular",
+        end_date: "2021-12-10T00:00:00.000Z",
+        hours: 80,
+        position_code: "invalidcourse",
+        start_date: "2020-12-10T00:00:00.000Z",
+        utorid: "potterh",
+        wage_chunks: [
+            {
+                end_date: "2020-12-31T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2020-12-10T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-06-30T00:00:00.000Z",
+                hours: 30,
+                rate: 50,
+                start_date: "2021-01-01T00:00:00.000Z",
+            },
+            {
+                end_date: "2021-12-10T00:00:00.000Z",
+                hours: 20,
+                rate: 50,
+                start_date: "2021-07-01T00:00:00.000Z",
+            },
+        ],
+    },
+];
+
+// ddah data to compare and compute diff
 export const initialApplicantDataForDdah = [
     {
         first_name: "Hanna",
@@ -383,3 +651,45 @@ export const initialDdahData = {
         },
     ],
 };
+
+export const modifiedDdahData = [
+    {
+        applicant: "potterh",
+        duties: [
+            {
+                description: "Initial training",
+                hours: 80,
+            },
+            {
+                description: "Marking the midterm",
+                hours: 10,
+            },
+            {
+                description: "Running tutorials",
+                hours: 20,
+            },
+        ],
+        position_code: "CSC135H1F",
+    },
+];
+
+export const modifiedDdahDataInvalidAssignment = [
+    {
+        applicant: "potterh",
+        duties: [
+            {
+                description: "Initial training",
+                hours: 80,
+            },
+            {
+                description: "Marking the midterm",
+                hours: 10,
+            },
+            {
+                description: "Running tutorials",
+                hours: 20,
+            },
+        ],
+        position_code: "invalid",
+    },
+];
